create server fn
